general component for documents
